requirements.txt
Flask==3.0.3
firebase-admin==6.5.0
python-dotenv==1.0.1
opencv-python==4.10.0.84
dlib==19.24.6
scipy==1.13.1
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1